Create remote-joystick
This code is used on ESP8266 module, a WEMOS D1 in Arduino Uno  form factor.

It is connected to a generic USB joystick encoder via a 5V relay. 

The WEMOS hosts a web page with a button, clicking the button activates the relay which presses a joystick button on the computer that is connected to the  generic USB joystick encoder.